requirements.txt
psycopg2-binary
pandas
spacy
seaborn
matplotlib
selenium